Move skynet.utils to skynet.dgpu submodule
